+++ site/src/hooks/index.ts
@@ -2,7 +2,6 @@ export * from "./useClickable"
 export * from "./useClickableTableRow"
 export * from "./useClipboard"
 export * from "./useFeatureVisibility"
-export * from "./useFilter"
 export * from "./useLocalStorage"
 export * from "./useMe"
 export * from "./useOrganizationId"
